Move Claim, NodeAux and Proof to types.go
